Refactor: Get rid of interfaces folder
